+++ admin/svelte/src/menu-queries-types.ts
@@ -3,7 +3,7 @@
  */
 
 export interface SelectOption {
-  value: string | number;
+  value: string;
   label: string;
 }